chore: combine promote commands
