+++ schemas/devops-promote.json
@@ -1,0 +1,28 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$ref": "#/definitions/PromoteResult",
+  "definitions": {
+    "PromoteResult": {
+      "type": "object",
+      "properties": {
+        "requestId": {
+          "type": "string"
+        },
+        "status": {
+          "type": "string"
+        },
+        "message": {
+          "type": "string"
+        },
+        "promotedWorkitemIds": {
+          "type": "array",
+          "items": {
+            "type": "string"
+          }
+        }
+      },
+      "required": ["requestId", "status", "message", "promotedWorkitemIds"],
+      "additionalProperties": false
+    }
+  }
+}
